Update the DSFR from 1.0 to 1.5 (#10220)
* chore: remove some spurious files

* npm: update the design system

and update the path to the new assets (the js/css folders are gone).

* footer: use the new footer from the DSFR

And remove a bunch of icons that are now provided by the design system
directly. This requires a PR for the GitHub icon[1] to be merged.

[1]: https://github.com/GouvernementFR/dsfr/pull/321

* footer: use the icons straight from the DSFR

* card: adapt the cards to the new design system

They have changed drastically from 1.0 to 1.5, which means we've had
to make a couple changes:

- they're slightly darker grey ;
- they have a border because it's nicer ;
- they use some custom padding and colour for the image because the
  white+grey variant we used previously is gone from the design
  system.

Also in this commit:

Previously we did a bunch of tricks to figure out the current
startups (`author.startups`) and the previous
one (`author.previously`) and then concat the two and show an extra
label if there were more than 5, and show all previous ones with an
inactive tag.

I take it on me this is absolutely non-essential and does not justify
the complexity of the previous code.

* chore: formatting

* brigade: adjust some of the old cards

Cards now have a `fr-card__content` container that provides the
adequate padding we rely on.

* startups: _openfisca_ is just openfisca

* npm: remove yarn.lock

We already have package-lock.json

* jobs: fix the job cards

Again some substantial modifications because we try and leverage the
new design system:

- default icons ;
- remove some unnecessary information ("friends", "open to juniors").

* latin: alumnus -> alumni

* communauté: fix the page

* head: don't include analytics if we're in development

* investigation: indentation

* job: enforce the h2 colour

the DSFR now enforces a colour for the H2s which made this one lose
the one in the `.section-color`.

* jobs: fix the tags

* jobs: fix the side section

* home: fix some buttons styling

* add newsletter in footer and remove from other pages (#10234)

* add newsletter in footer and remove from other pages

* remove useless classes and div

* intra: remove a spurious link class

* jobs: remove a reference to a dead CSS file

* js: adjust the member search to the new cards

as with the regular HTML, drop the inactive/active tag logic for
previously, current startups. Note to us: we should really use some
form of JS library to avoid the duplication of markup/logic that's
incredibly painful to maintain.

* startups: lotocar is now comobi

* jobs: rewrite the job CTAs without the card layout

Restore the previous `.section-blue h3:not(.fr-card-title)` because
that section blue is used elsewhere with some cards in it and it
messes them up. Avoid the cards layout in the jobs section sidebar to
get rid of the problem.

* intra: fix some of the smaller cards

* home: use cover for the object-fit of the showcase startups

* remove default icon from the DSFR on target blank for newsletter form

* harmonize and fix grey card - include card template in intra view

* fix hybrid component between card and tile on brigade page

* harmonize forgotten white button on purple background

* fix filter on hover on jobs page using button instead of anchor tags

* use appropriate component for summary

* remove class for external link icon because its included in  DSFR

* remove unwanted space to push and try deploy on scalingo

* change aspect-ratio on rounded image card to avoid bad ratio on Chrome

* fix startups search display

* fix cards on approche/formation page

* add bottom margin under jobs card on startup page

* remove target blank on card title in search-startup.js

* add target blank on card community if link and fix link on annuaire card after filter

Co-authored-by: Lisa Durand <lisa.c.durand@gmail.com>