ci: remove our use of Python
No one take this personally please :)

We don't need to check our YAML syntax because Jekyll and Kwalify will
both load our files and error if needed.

Regarding the actual semantics of our metadata, we already have the
`schema/authors.yml` and `schema/startups.yml` which should be used
for that purpose. I've updated the `end` attribute to be a date and fixed
the few places where it wasn't a date. `phases` is already required,
we don't need to check that again.

There are no startups with mismatched start and end date (i.e start >
end) but if we really want to check on that let's use the new BetaDsl
instead (lib/beta_dsl/startup.rb) rather than a whole new language and
a bunch of hacks. Here:

BetaDsl::Startup.all.count do |startup|
  startup.phases.any? do |phase|
    phase['start'] && phase['end'] && phase['start'] > phase['end']
  end
end

does the job (and yields 0).