fast: add the FAST 12 laureates and make it dynamic
Add a 'fast' property on the startups pages/objects to indicate a
startup that was selected for that promotion, and add a dynamic
partial that is triggered via a 'new-fast' property on the actual
page.

This allows the FAST team to make the future updates themselves, for example
with the next edition:

1. add a new 13_fast.md page that has a `new-fast: true` property ;
2. go through the startups that won and add a the fast attributes:

```yml
fast:
  promotion: 13
  montant: 10000
```

And then it just works™

fast: add the FAST 12 laureates and make it dynamic

Add the FAST information directly on the startup Frontmatter (i.e the
YAML data) and make the FAST 12 page use it via a new dynamic partial
that's included if the "new-fast" page attribute is true.

This allows This means that in the future the FAST team can do it all themselves,
for example with the next edition:

1. add a new 13_fast.md page that has a `new-fast: true` property ;
2. go through the startups that won and add a `fast` YML prop with the
`promotion`, `amount`, `ambition` and `impact` properties in there.

And then it just works™