Fix/api size (#20819)
* config: don't transpile the lib files

* lib: don't mutate the data objects we use to instantiate Member

This has cost me a day of work but it is incredibly simple: because we
initialize our Beta::Member class with the Jekyll::Document hash, it
means doing

1. @startups = value from the hash
2. def active_startups; @startups.concat(...); end

means we're actually modifying the value from the hash since it's just
a reference being passed around. Jekyll maintains that `doc.data`
reference to build up the template but that references keeps growing
with every call to the methods `active_startups` or `legacy_startup`.

My bad! I thought `Array#concat` returned a new array...