plugins: ensure the community plugin considers all missions
The code used on `/annuaire` uses the Community plugin which allows
querying everybody that's active, and everybody that's not (i.e:
alumni).

To figure out if they are alumni, it was grabbing the last
mission (missions.last) and comparing the end date with today. That's
fair but it turns out that our member files are, obviously, not always
sorted chronologically.

Some of them have an early mission (early in the array) that's still
active, but with the previous logic it's only the last mission
considered hence why they don't appaear in the current members tab.

So instead use the glorious DSL (lib/member.rb) again to wrap the data
and query for any `active_missions` which removes some duplicate logic
and the display bug as well.